Update URL for Dart rules (#87)
The official branch of the Dart rules for Bazel was deprecated and
archived in 2018. See: https://github.com/dart-archive/rules_dart

Unfortunately, there is currently no officially-maintained open source
repo of Bazel rules for Dart and the ones we previously had have been
deprecated and archived.

This updates to what I believe to be the only maintained (on an ad-hoc
basis) set of Bazel rules for Dart of which I'm aware.
cbracken/rules_dart is where I developed the original rules for use at
Google and formed the foundation on which we built dart-lang/rules_dart.

The Dart team has moved away from Bazel support in order to focus on a
lighter-weight home-grown build system which is a better fit for most
smaller Dart repos. Discussion can be found here:
https://groups.google.com/a/dartlang.org/g/misc/c/KshWeq5kuzI

While I can't promise these will continue to track the bleeding edge of
Dart SDK and Bazel releases, I have been maintaining them since 2016 and
am happy to take PRs.